requirements1.txt
OpenCV
NumPy
Pillow
MySQL-Connector
Pandas
Tkinter
Haar Cascade
